refactor: add ui dir
