+++ packages/cats/src/components/templates/FooterLogo/FooterLogo.tsx
