+++ packages/cats/src/components/templates/FooterLogo/index.ts
