+++ packages/cats/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from "next/app";
 // import Script from "next/script";
 
 import "tailwindcss/tailwind.css";
-import { Seo } from "@/cats/components/Seo";
+import { Seo } from "@/cats/components/templates/Seo";
 
 const CustomApp = ({ Component, pageProps }: AppProps): JSX.Element => {
   return (
